Clarifications to SPV_KHR_cooperative_matrix
- Change the description of 'Memory Operand' operands to
load/store instructions to make it clear that multiple words can
be used (in line with the language used for OpLoad/OpStore).
- Require all instructions that have a cooperative matrix object as
an operand or produce a cooperative matrix object result only be
executed in uniform control flow within the scope of the matrix type.
- Require the constituent provided to OpCompositeConstruct to be
dynamically uniform within the scope of the cooperative matrix.
- Require the operands to arithmetic and conversion instructions
operating on cooperative matrix objects to be dynamically uniform
within the scope of the cooperative matrix.
- Clarify when _Stride_ and _Memory Operand_ operands can be provided.
- Clarify when referring to cooperative matrix types or objects/values
of that type in the description of conversion instructions.
- Clarify that elementwise access to cooperative matrix objects can be
performed in non-uniform control flow.


Fixes #208
Fixes #211

Signed-off-by: Kevin Petit <kevin.petit@arm.com>
Change-Id: Ia2b429e4da410ba51ac6da9379c50875bb354001