Allow DebugTypePointer BaseType argument substitution by DebugInfoNone
This update addresses the DebugTypePointer instruction, which represents pointer types in the SPIR-V NonSemantic.Shader.DebugInfo.100 standard.
Currently, the instruction requires a defined Base Type argument to comply with the standard.
This requirement complicates the implementation of pointers to abstract types,
such as void * in C, which represents type erasure, since these types do not have an available or defined Base Type.

To accommodate such scenarios, the standard includes the DebugInfoNone instruction.
This amendment allows for the substitution of the concrete Base Type argument in DebugTypePointer with DebugInfoNone,
thereby simplifying the implementation process for abstract pointer types.